Convert `http` links to `https` links
Several `http` href links were converted to `https`. All of these links
are to servo.org and subdomains, or to rust-lang.org, which support
https.

The links to the starters page were updated to
`starters.servo.org`.

The exceptions are urls in license files, and two urls in bhtml-newtab:
`arstechnica.com`, and the jquery cdn. The last one was exluded from
this PR, because it modifies something in the browser, and should be
replaced with care (and testing?)